+++ bootstrap.js
@@ -2,6 +2,7 @@ import App from "@bejibun/app";
 import Database from "@bejibun/database";
 import BaseModel from "./bases/BaseModel";
 import Kernel from "./decorators/Kernel";
+import "./globals/index";
 import NamespaceLoader from "./loader/NamespaceLoader";
 BaseModel.knex(Database.knex());
 Kernel.registerDecorator();