+++ config/performance.js
@@ -1,0 +1,7 @@
+const config = {
+    middlewares: {
+        limiter: true,
+        maintenance: true
+    }
+};
+export default config;