+++ server.d.ts
@@ -2,5 +2,6 @@ export default class Server {
     private get exceptionHandler();
     private get apiRoutes();
     private get webRoutes();
+    private get performance();
     run(): Promise<void>;
 }