+++ src/server.ts
@@ -2,6 +2,8 @@ import type {RouterGroup} from "@/types";
 import App from "@bejibun/app";
 import Logger from "@bejibun/logger";
 import {defineValue} from "@bejibun/utils";
+import fs from "fs";
+import PerformanceConfig from "@/config/performance";
 import RuntimeException from "@/exceptions/RuntimeException";
 import Router from "@/facades/Router";
 import MaintenanceMiddleware from "@/middlewares/MaintenanceMiddleware";
@@ -39,6 +41,17 @@ export default class Server {
         } catch (error: any) {
             throw new RuntimeException(`Missing web file on routes directory [${webRoutesPath}].`, null, error.message);
         }
+    }
+
+    private get performance(): Record<string, any> {
+        const configPath: string = App.Path.configPath("performance.ts");
+
+        let config: any;
+
+        if (fs.existsSync(configPath)) config = require(configPath).default;
+        else config = PerformanceConfig;
+
+        return config;
     }
 
     public async run(): Promise<void> {
@@ -85,8 +98,13 @@ export default class Server {
             paths: paths
         }, null, 2));
 
+        const middlewares: Array<any> = [];
+
+        if (this.performance.middlewares.limiter) middlewares.push(new RateLimiterMiddleware());
+        if (this.performance.middlewares.maintenance) middlewares.push(new MaintenanceMiddleware());
+
         const server = Bun.serve({
-            development: Bun.env.NODE_ENV !== "production" && {
+            development: defineValue(Bun.env.APP_ENV, "development") !== "production" && {
                 // Enable browser hot reloading in development
                 hmr: true,
 
@@ -102,10 +120,7 @@ export default class Server {
                 "/": require(App.Path.publicPath("index.html")),
                 "/apis": require(App.Path.publicPath("apis.html")),
 
-                ...Object.assign({}, ...defineValue(Router.middleware(
-                    new MaintenanceMiddleware(),
-                    new RateLimiterMiddleware()
-                ).group([
+                ...Object.assign({}, ...defineValue(Router.middleware(...middlewares).group([
                     Router.namespace("app/exceptions").any("/*", "Handler@route"),
 
                     apiRoutes,