+++ types/index.d.ts
@@ -1,4 +1,5 @@
 export * from "../types/decorator";
+export * from "../types/global";
 export * from "../types/middleware";
 export * from "../types/router";
 export * from "../types/schedule";